requirements.txt
asgiref
Django
django-filter
djangorestframework
pytz
sqlparse
drf-yasg
psycopg2-binary